+++ perl/FSM/Composition/RealizedInstance.pm
@@ -3,10 +3,9 @@ package FSM::Composition::RealizedInstance;
 use v5.20;
 use strict;
 use warnings;
-use Carp qw(confess);
 use feature qw(signatures);
 no warnings 'experimental::signatures';
-use FSM::IR::StructuralRTLIR::ConnectionExpr qw(signal_ref_expr expr_signal_name);
+use FSM::IR::StructuralRTLIR::ConnectionExpr qw(normalized_binding);
 
 sub new ($class, %args) {
     return bless {
@@ -32,48 +31,7 @@ sub hdl_code ($self) { return $self->{hdl_code} }
 
 sub _normalize_port_bindings ($bindings) {
     return [] unless ref($bindings) eq 'ARRAY';
-    return [ map { _normalize_port_binding($_) } @$bindings ];
-}
-
-sub _normalize_port_binding ($binding) {
-    confess "RealizedInstance port bindings must be hash entries"
-        unless ref($binding) eq 'HASH';
-
-    my $port_name = $binding->{port_name};
-    my $signal_name = $binding->{signal_name};
-    my $connection_expr = _clone($binding->{connection_expr});
-
-    if (!defined($signal_name) || !length($signal_name)) {
-        $signal_name = expr_signal_name($connection_expr);
-    }
-
-    if ((!ref($connection_expr) || ref($connection_expr) ne 'HASH')
-        && defined($signal_name) && length($signal_name))
-    {
-        $connection_expr = signal_ref_expr($signal_name);
-    }
-
-    return {
-        port_name => $port_name,
-        signal_name => $signal_name,
-        connection_expr => $connection_expr,
-    };
-}
-
-sub _clone ($value) {
-    return undef unless defined $value;
-
-    if (ref($value) eq 'HASH') {
-        return {
-            map { $_ => _clone($value->{$_}) } sort keys %$value
-        };
-    }
-
-    if (ref($value) eq 'ARRAY') {
-        return [ map { _clone($_) } @$value ];
-    }
-
-    return $value;
+    return [ map { normalized_binding($_) } @$bindings ];
 }
 
 1;